chore: refactor headerbar
