+++ packages/components/src/HeaderBar/primitives/AppNotification.component.js
@@ -3,20 +3,19 @@ import { useMemo } from 'react';
 import omit from 'lodash/omit';
 import { useTranslation } from 'react-i18next';
 
-import Inject from '../Inject';
-import Action from '../Actions/Action';
-import ActionIntercom from '../ActionIntercom';
-import ActionDropdown from '../Actions/ActionDropdown';
-import Typeahead from '../Typeahead';
-import I18N_DOMAIN_COMPONENTS from '../constants';
-import { getTheme } from '../theme';
-import AppSwitcher from '../AppSwitcher';
+import Inject from '../../Inject';
+import { Action } from '../../Actions';
+import ActionIntercom from '../../ActionIntercom';
+import Typeahead from '../../Typeahead';
+import I18N_DOMAIN_COMPONENTS from '../../constants';
+import { getTheme } from '../../theme';
+import AppSwitcher from '../../AppSwitcher';
 
-import headerBarCssModule from './HeaderBar.module.scss';
+import headerBarCssModule from '../HeaderBar.module.scss';
 
 const theme = getTheme(headerBarCssModule);
 
-function AppNotification({ getComponent, hasUnread, t, ...props }) {
+export function AppNotification({ getComponent, hasUnread, t, ...props }) {
 	const className = theme('tc-header-bar-action', 'separated');
 
 	let icon;
@@ -50,216 +49,9 @@ function AppNotification({ getComponent, hasUnread, t, ...props }) {
 	);
 }
 
-function Intercom({ id, config, tooltipPlacement }) {
-	return (
-		<ActionIntercom
-			className={theme('tc-header-bar-intercom-default-component', 'btn', 'btn-link')}
-			id={id}
-			config={useMemo(() => ({ ...config, vertical_padding: 70 }), [config])}
-			tooltipPlacement={tooltipPlacement}
-		/>
-	);
-}
-
-function GenericAction({ getComponent, ...props }) {
-	const global = {
-		bsStyle: 'link',
-		tooltipPlacement: 'bottom',
-		...props,
-	};
-	const className = theme('tc-header-bar-action', 'tc-header-bar-generic-action', 'separated');
-	const Renderers = Inject.getAll(getComponent, { Action });
-
-	return (
-		<li role="presentation" className={className}>
-			<Renderers.Action {...global} />
-		</li>
-	);
-}
-
-function HeaderBar(props) {
-	const { t } = useTranslation(I18N_DOMAIN_COMPONENTS);
-
-	const Components = Inject.getAll(props.getComponent, {
-		Logo,
-		Environment,
-		CallToAction,
-		GenericAction,
-		Search,
-		User,
-		Information,
-		Help,
-		AppNotification,
-		Intercom,
-	});
-
-	if (props.brand && props.products) {
-		console.warn('Deprecated: use @talend/ui-ee/AppSwitcher');
-	}
-
-	const AppSwitcherComponent =
-		props.AppSwitcher || Inject.get(props.getComponent, 'AppSwitcher', AppSwitcher);
-
-	let notificationCenter;
-	const { NotificationCenter } = props;
-	if (NotificationCenter) {
-		notificationCenter = <NotificationCenter />;
-	} else if (props.notification) {
-		console.warn('Deprecated: use @talend/notification-center');
-		notificationCenter = (
-			<Components.AppNotification getComponent={props.getComponent} {...props.notification} t={t} />
-		);
-	}
-
-	let intercom;
-	const { Intercom: CustomIntercom } = props;
-	if (CustomIntercom) {
-		intercom = <CustomIntercom />;
-	} else if (props.intercom) {
-		console.warn('Deprecated: use @talend/ui-ee/Intercom');
-		intercom = <Components.Intercom getComponent={props.getComponent} {...props.intercom} />;
-	}
-
-	return (
-		<nav className={theme('tc-header-bar', 'navbar')}>
-			<ul className={theme('tc-header-bar-actions', 'navbar-nav')}>
-				{props.logo && <Components.Logo getComponent={props.getComponent} {...props.logo} t={t} />}
-				<AppSwitcherComponent {...props.brand} {...props.products} isSeparated={!!props.env} />
-				{props.env && <Components.Environment getComponent={props.getComponent} {...props.env} />}
-			</ul>
-			<ul className={theme('tc-header-bar-actions', 'navbar-nav', 'right')}>
-				{props.genericAction && (
-					<Components.GenericAction getComponent={props.getComponent} {...props.genericAction} />
-				)}
-				{props.callToAction && (
-					<Components.CallToAction getComponent={props.getComponent} {...props.callToAction} />
-				)}
-				{props.search && <Components.Search getComponent={props.getComponent} {...props.search} />}
-				{notificationCenter && (
-					<li
-						role="presentation"
-						className={theme(
-							'tc-header-bar-notification-center',
-							'tc-header-bar-action',
-							'separated',
-						)}
-					>
-						{notificationCenter}
-					</li>
-				)}
-				{intercom && (
-					<li
-						role="presentation"
-						className={theme('tc-header-bar-intercom', 'tc-header-bar-action', 'separated')}
-					>
-						{intercom}
-					</li>
-				)}
-				{props.help && <Components.Help getComponent={props.getComponent} {...props.help} t={t} />}
-				{!props.user && props.information && (
-					<Components.Information getComponent={props.getComponent} {...props.information} t={t} />
-				)}
-				{props.user && <Components.User getComponent={props.getComponent} {...props.user} t={t} />}
-			</ul>
-		</nav>
-	);
-}
-
-HeaderBar.Logo = Logo;
-HeaderBar.Environment = Environment;
-HeaderBar.CallToAction = CallToAction;
-HeaderBar.Search = Search;
-HeaderBar.GenericAction = GenericAction;
-HeaderBar.Help = Help;
-HeaderBar.Information = Information;
-HeaderBar.User = User;
-HeaderBar.displayName = 'HeaderBar';
-
-if (process.env.NODE_ENV !== 'production') {
-	Search.propTypes = {
-		...Typeahead.propTypes,
-		renderers: PropTypes.shape({
-			Typeahead: PropTypes.func,
-		}),
-	};
-
-	GenericAction.propTypes = {
-		getComponent: PropTypes.func,
-		renders: PropTypes.shape({
-			Action: PropTypes.func,
-		}),
-	};
-
-	Help.propTypes = {
-		getComponent: PropTypes.func,
-		renderers: PropTypes.shape({
-			ActionSplitDropdown: PropTypes.func,
-			Action: PropTypes.func,
-		}),
-		t: PropTypes.func.isRequired,
-	};
-
-	Information.propTypes = {
-		getComponent: PropTypes.func,
-		items: PropTypes.array,
-		renderers: PropTypes.shape({
-			ActionSplitDropdown: PropTypes.func,
-			Action: PropTypes.func,
-		}),
-		t: PropTypes.func.isRequired,
-	};
-
-	User.propTypes = {
-		firstName: PropTypes.string,
-		lastName: PropTypes.string,
-		getComponent: PropTypes.func,
-		name: PropTypes.string,
-		renderers: PropTypes.shape({ ActionDropdown: PropTypes.func }),
-		t: PropTypes.func,
-	};
-
-	AppNotification.propTypes = {
-		getComponent: PropTypes.func,
-		hasUnread: PropTypes.bool,
-		renderers: PropTypes.shape({ Action: PropTypes.func }),
-		t: PropTypes.func.isRequired,
-	};
-
-	Intercom.propTypes = {
-		id: PropTypes.string.isRequired,
-		config: PropTypes.object.isRequired,
-		tooltipPlacement: PropTypes.string,
-	};
-
-	HeaderBar.propTypes = {
-		AppSwitcher: PropTypes.func,
-		Intercom: PropTypes.func,
-		NotificationCenter: PropTypes.func,
-		logo: PropTypes.shape(omit(Logo.propTypes, 't')),
-		brand: PropTypes.shape({
-			isSeparated: PropTypes.bool,
-			renderers: PropTypes.shape({
-				Action: PropTypes.func,
-			}),
-			icon: PropTypes.string,
-			iconUrl: PropTypes.string,
-		}),
-		env: PropTypes.shape(Environment.propTypes),
-		callToAction: PropTypes.shape(CallToAction.propTypes),
-		genericAction: PropTypes.shape(GenericAction.propTypes),
-		search: PropTypes.shape(Search.propTypes),
-		help: PropTypes.shape(omit(Help.propTypes, 't')),
-		information: PropTypes.shape(omit(Information.propTypes, 't')),
-		intercom: PropTypes.shape(Intercom.propTypes),
-		user: PropTypes.shape(User.propTypes),
-		notification: PropTypes.shape(omit(AppNotification.propTypes, 't')),
-		products: PropTypes.shape({
-			items: PropTypes.array,
-			onSelect: PropTypes.func,
-		}),
-		getComponent: PropTypes.func,
-		t: PropTypes.func,
-	};
-}
-
-export default HeaderBar;
+AppNotification.propTypes = {
+	getComponent: PropTypes.func,
+	hasUnread: PropTypes.bool,
+	renderers: PropTypes.shape({ Action: PropTypes.func }),
+	t: PropTypes.func.isRequired,
+};
